requirements.txt
# =============================================================================
# Python 3.14+ (use environment.yml for conda, or pyenv/uv install 3.14).
# =============================================================================

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.25
asyncpg>=0.29.0
greenlet>=3.0.0
alembic>=1.13.0

# =============================================================================
# Redis & Background Tasks
# =============================================================================
redis>=5.0.0
taskiq>=0.11.0
taskiq-redis>=1.0.0

# =============================================================================
# Event Bus
# =============================================================================
faststream[redis]>=0.5.0

# =============================================================================
# HTTP Client
# =============================================================================
httpx>=0.26.0

# =============================================================================
# Security
# =============================================================================
python-jose[cryptography]>=3.3.0
cryptography>=43.0
python-multipart>=0.0.6
email-validator>=2.0

# =============================================================================
# Email
# =============================================================================
resend>=2.0

# =============================================================================
# SSE (Server-Sent Events)
# =============================================================================
sse-starlette>=2.0

# =============================================================================
# Logging & Observability
# =============================================================================
structlog>=24.1.0
# --- Tracing (hooks only — disabled by default) ---
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
# --- Metrics (hooks only — disabled by default) ---
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.20.0

# =============================================================================
# Resilience
# =============================================================================
aiobreaker>=1.2.0
tenacity>=9.0.0
uvloop>=0.21.0; sys_platform != "win32"

# =============================================================================
# Configuration
# =============================================================================
pyyaml>=6.0.1
python-dotenv>=1.0.0

# =============================================================================
# CLI
# =============================================================================
click>=8.1.0
rich>=13.7.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
respx>=0.22

# =============================================================================
# Code Quality
# =============================================================================
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
